Extract shared code
